PR comments - move `NullHtmlEncoder` to `System.Text.Encodings.Web` namespace - remove parameter name in generated code
