+++ src/Microsoft.AspNet.Razor.Runtime/TagHelpers/NullHtmlEncoder.cs
@@ -1,9 +1,11 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System;
 using System.IO;
+using System.Text.Encodings.Web;
 
-namespace System.Text.Encodings.Web
+namespace Microsoft.AspNet.Razor.TagHelpers
 {
     /// <summary>
     /// A <see cref="HtmlEncoder"/> that does not encode. Should not be used when writing directly to a response
