+++ test/Microsoft.AspNet.Razor.Test/TestFiles/CodeGenerator/Output/NestedScriptTagTagHelpers.cs
@@ -54,7 +54,7 @@ namespace TestOutput
                 __tagHelperExecutionContext.Add(__TestNamespace_InputTagHelper);
                 __TestNamespace_InputTagHelper2 = CreateTagHelper<global::TestNamespace.InputTagHelper2>();
                 __tagHelperExecutionContext.Add(__TestNamespace_InputTagHelper2);
-                StartTagHelperWritingScope(encoder: null);
+                StartTagHelperWritingScope(null);
                 WriteLiteral("2000 + ");
 #line 8 "NestedScriptTagTagHelpers.cshtml"
                                             Write(ViewBag.DefaultInterval);
