+++ test/Microsoft.AspNet.Razor.Test/TestFiles/CodeGenerator/Output/PrefixedAttributeTagHelpers.cs
@@ -172,7 +172,7 @@ __TestNamespace_InputTagHelper1.IntDictionaryProperty["salt"] = 37;
             __TestNamespace_InputTagHelper1.StringDictionaryProperty["paprika"] = "another string";
             __tagHelperExecutionContext.AddTagHelperAttribute("string-prefix-paprika", __TestNamespace_InputTagHelper1.StringDictionaryProperty["paprika"]);
             __TestNamespace_InputTagHelper2.StringDictionaryProperty["paprika"] = __TestNamespace_InputTagHelper1.StringDictionaryProperty["paprika"];
-            StartTagHelperWritingScope(encoder: null);
+            StartTagHelperWritingScope(null);
             WriteLiteral("literate ");
 #line 21 "PrefixedAttributeTagHelpers.cshtml"
                              WriteLiteral(literate);
